+++ packages/common/src/types/featureFlags.ts
@@ -24,11 +24,6 @@ export enum FeatureFlags {
     /**/
     CustomSQLEnabled = 'custom-sql-enabled',
 
-    /**/
-    PuppeteerScrollElementIntoView = 'puppeteer-scroll-element-into-view',
-    PuppeteerSetViewportDynamically = 'puppeteer-set-viewport-dynamically',
-    PuppeteerDisconnectBrowser = 'puppeteer-disconnect-browser',
-
     /* Shows the two-stage login flow */
     newLoginEnabled = 'new-login-enabled',