+++ packages/common/src/utils/item.ts
@@ -4,7 +4,9 @@ import {
     fieldId,
     isDimension,
     isField,
+    isTableCalculation,
     MetricType,
+    TableCalculationType,
     type CompiledDimension,
     type CustomDimension,
     type Dimension,
@@ -19,8 +21,11 @@ import {
     type AdditionalMetric,
 } from '../types/metricQuery';
 
-export const isNumericType = (type: DimensionType | MetricType) => {
+export const isNumericType = (
+    type: DimensionType | MetricType | TableCalculationType,
+) => {
     const numericTypes = [
+        TableCalculationType.NUMBER,
         DimensionType.NUMBER,
         MetricType.NUMBER,
         MetricType.PERCENTILE,
@@ -47,8 +52,10 @@ export const isNumericItem = (
         return false;
     }
     if (isCustomDimension(item)) return false;
-    if (isField(item) || isAdditionalMetric(item)) {
-        return isNumericType(item.type as DimensionType | MetricType);
+    if (isField(item) || isAdditionalMetric(item) || isTableCalculation(item)) {
+        return isNumericType(
+            item.type as DimensionType | MetricType | TableCalculationType,
+        );
     }
     return true;
 };