+++ packages/frontend/src/hooks/echarts/useEchartsCartesianConfig.ts
@@ -24,10 +24,13 @@ import {
     TimeFrames,
     type ApiQueryResults,
     type CartesianChart,
+    type CustomDimension,
+    type Field,
     type ItemsMap,
     type PivotReference,
     type ResultRow,
     type Series,
+    type TableCalculation,
 } from '@lightdash/common';
 import dayjs from 'dayjs';
 import {
@@ -101,6 +104,14 @@ const getAxisTypeFromField = (item?: ItemsMap[string]): string => {
             case MetricType.DATE:
             case TableCalculationType.DATE:
             case TableCalculationType.TIMESTAMP:
+                // Use categorical axis for weeks only. Echarts handles the
+                // other time frames well with a time axis
+                if (
+                    'timeInterval' in item &&
+                    item.timeInterval === TimeFrames.WEEK
+                ) {
+                    return 'category';
+                }
                 return 'time';
             default: {
                 return 'category';
@@ -802,6 +813,33 @@ const getLongestLabel = ({
     );
 };
 
+const getWeekAxisConfig = (
+    axisId?: string,
+    axisField?: Field | TableCalculation | CustomDimension,
+    rows?: ResultRow[],
+) => {
+    if (!axisId || !rows || !axisField) return {};
+    if (
+        'timeInterval' in axisField &&
+        axisField.timeInterval === TimeFrames.WEEK
+    ) {
+        const [minX, maxX] = getMinAndMaxValues([axisId], rows || []);
+        const continuousWeekRange = [dayjs.utc(minX).format()];
+        let nextDate = dayjs.utc(minX);
+        while (nextDate.isBefore(dayjs(maxX))) {
+            nextDate = nextDate.add(1, 'week');
+            continuousWeekRange.push(nextDate.format());
+        }
+        continuousWeekRange.push(dayjs.utc(maxX).format());
+        return {
+            data: continuousWeekRange,
+            axisTick: { alignWithLabel: true, interval: 0 },
+        };
+    } else {
+        return {};
+    }
+};
+
 const getEchartAxes = ({
     itemsMap,
     validCartesianConfig,
@@ -931,6 +969,7 @@ const getEchartAxes = ({
             // This is to ensure the value is correctly formatted on some types
             switch (axisItem.timeInterval) {
                 case TimeFrames.WEEK_NUM:
+                case TimeFrames.WEEK:
                     axisConfig.axisLabel = {
                         formatter: (value: any) => {
                             return formatItemValue(axisItem, value, false);
@@ -962,6 +1001,7 @@ const getEchartAxes = ({
                 (longestLabelWidth || 0) * Math.sin(rotateRadians);
             axisConfig.axisLabel = axisConfig.axisLabel || {};
             axisConfig.axisLabel.rotate = rotate;
+            axisConfig.axisLabel.margin = 12;
             axisConfig.nameGap = oppositeSide + 15;
         }
         return axisConfig;
@@ -1062,6 +1102,28 @@ const getEchartAxes = ({
         validCartesianConfig.eChartsConfig.series,
         itemsMap,
     );
+
+    const bottomAxisExtraConfig = getWeekAxisConfig(
+        bottomAxisXId,
+        bottomAxisXField,
+        resultsData?.rows,
+    );
+    const topAxisExtraConfig = getWeekAxisConfig(
+        topAxisXId,
+        topAxisXField,
+        resultsData?.rows,
+    );
+    const rightAxisExtraConfig = getWeekAxisConfig(
+        rightAxisYId,
+        rightAxisYField,
+        resultsData?.rows,
+    );
+    const leftAxisExtraConfig = getWeekAxisConfig(
+        leftAxisYId,
+        leftAxisYField,
+        resultsData?.rows,
+    );
+
     return {
         xAxis: [
             {
@@ -1115,6 +1177,7 @@ const getEchartAxes = ({
                         : showGridX,
                 },
                 inverse: !!xAxisConfiguration?.[0].inverse,
+                ...bottomAxisExtraConfig,
             },
             {
                 type: topAxisType,
@@ -1155,6 +1218,7 @@ const getEchartAxes = ({
                 splitLine: {
                     show: isAxisTheSameForAllSeries,
                 },
+                ...topAxisExtraConfig,
             },
         ],
         yAxis: [
@@ -1206,6 +1270,7 @@ const getEchartAxes = ({
                         : showGridY,
                 },
                 inverse: !!yAxisConfiguration?.[0].inverse,
+                ...leftAxisExtraConfig,
             },
             {
                 type: rightAxisType,
@@ -1250,6 +1315,7 @@ const getEchartAxes = ({
                 splitLine: {
                     show: isAxisTheSameForAllSeries,
                 },
+                ...rightAxisExtraConfig,
             },
         ],
     };